Reworked TerminalPixel.cs
Now holds a BackgroundColor and Foreground color value instead of just a single color for the background.

PixelContent is also now stored on the individual TerminalPixel instead of in a static dictionary that is fetched from it. This was done in concern to memory performance with extremely large amounts of pixel contents, especially since it was never cleaned during the lifetime of the application (leading to overflows..)